Move `HostName` to `mreg_cli.api.fields`
